split up utils
