+++ tests/test_utils.py
@@ -1,16 +1,11 @@
-# test python utilities
 import geopandas as gpd
 import pandas as pd
 import pytest
 
 from python.utils import (
-    NYC_PROJECTION,
-    WKT_PROJECTION,
-    convert_to_geodata,
     load_data_file,
     load_geodata_url,
     load_shapefile,
-    reporject_geometry,
 )
 
 TEST_DATA_PATH = "tests/test_data"
@@ -52,16 +47,3 @@ def test_load_shapefile_large():
     geography = load_shapefile("./.data/dev_db", "housing.shp.zip")
     assert len(geography) > 1
 
-
-def test_convert_to_geodata():
-    data = load_data_file(filepath=f"{TEST_DATA_PATH}/census_counties_nyc.csv")
-    data = convert_to_geodata(data)
-    assert isinstance(data, gpd.GeoDataFrame)
-    assert data.crs == WKT_PROJECTION
-
-
-def test_reporject_geometry():
-    data = load_data_file(filepath=f"{TEST_DATA_PATH}/census_counties_nyc.csv")
-    data = convert_to_geodata(data)
-    data = reporject_geometry(data, WKT_PROJECTION, NYC_PROJECTION)
-    assert data.crs == NYC_PROJECTION
